docs: Add Raycast example for flame_forge2d (#1253)
Adds a new sample in flame_forge2d for Raycast. It contains a sample implementation for finding nearest and farthest fixtures along the casted ray. Here is how it works, two rays (red and blue) are fired from same starting point with mouse position as the target. Red ray always stops at the nearest fixture and blue ray always stop at the farthest fixture. Paint of fixtures are changed depending on which ray hits it. If both rays hit the same fixture, it turns yellow.