feat(effects): Added SpeedEffectController (#1260)
This adds SpeedEffectController -- new kind of effect controller that can be used to create effects that proceed with a predefined speed. In addition, the factory constructor EffectController() now has 2 more parameters speed and reverseSpeed for easier creation of speed-based effects. The argument duration is now optional.

The base EffectController class now has onMount() callback, which is invoked when the effect controller is attached to an Effect.