feat: Adding bloc getter to FlameBlocListenable mixin (#1732)
This PR adds a getter to the FlameBlocListenable mixin. This getter will allow to use FlameBlocListenable as a standalone mixin to react to new states via onNewState(), access the current state, and also add events to the listenable Bloc via the bloc property.
Currently a workaround to give components this functionality is to use FlameBlocListenable and FlameBlocReader in parallel.