feat: Added TextFormatter classes (#1720)
    A TextFormatter is a class that can take a string of text and convert it into a TextElement.
    A TextElement is an object that encapsulates a text string prepared for rendering. This object knows its own dimensions (layout), and is able to render itself on a canvas.

A TextFormatter is different from the current TextRenderer in one crucial respect: whereas TextRenderer takes a string of text and draws it onto the canvas directly (performing the layout and measurement internally), the TextFormatter creates an object that encapsulates the information necessary for rendering.

Thus, TextFormatter's output can be used to prepare the text once, and then render multiple times. In addition, since each TextElement knows its own layout, -- these individual layouts can be collected, manipulated, and arranged into a larger text body necessary for rich text rendering.

In addition, this PR:

    Implements a debug version of the standard text renderer, which draws rectangles instead of text;
    Re-enables the skipped golden test for the TextBoxComponent (hopefully it'll work this time);
    Refactors existing TextRenderers to use the new TextFormatters under the hood.

This PR is a WIP for the Rich Text functionality (#1627).