docs: Added ability to create infobox-style widgets in the documentation (#1771)
Additional mode "infobox" allows Flutter widgets to be inserted into the generated documentation site as wikipedia-style infoboxes floating on the right. This is useful for presenting content to the user up-front without sacrificing the flow of the exposition. For example, currently the Tap Events demo is "hidden" in the middle of a long page.

We believe that presenting the content in this way would make the user more excited to read about it.