feat: PositionComponent now has a built-in Decorator (#1846)
This PR adds the decorator property to PositionComponent, and makes decorators chain-able. That is, decorators function as linked lists, and a new decorator can be appended to the end of such list, and later removed.

The PositionComponent starts with the default decorator: the Transform2DDecorator, which encapsulates the functionality which previously resided within the PositionComponent.renderTree. This allows to:

apply/remove additional visual effects via .decorator.addLast()/.removeLast();

replace the Transform2DDecorator with another one -- for example supporting skew, or isometric coordinates, or 3d rotations.

HasDecorator mixin can no longer be applied to a PositionComponent because it already has the decorator "built in".

The decorators Shadow3DDecorator and Rotate3DDecorator, when added to the PositionComponent, will now apply after the coordinate transform of the PositionComponent, as desired. This means that, for example, the shadow's base and rotation pivot point can be specified using the local coordinates of the component.