feat: RecycledQueue now supports modification during iteration (#1884)
Added ability to add/remove elements into the RecycledQueue while iterating. This is needed for when the RecycledQueue will be used for lifecycle management, and there could be modifications of the queue invoked from onMount or onRemove callbacks.

The only restriction is that there can only be one iterator over the queue at a time.