tests: Clean up Component tests (#1896)
Going through some of the component-related tests and performing general cleanup:

Making sure the tests are in proper files (e.g. if a test tests the position of a PositionComponent, then it should be in the position_component_test.dart file);

Use testWithFlameGame / testWithGame where appropriate;

Some test names / variables names renamed for clarity;

Make sure the main() function is at the top of the test file;

Added test groups where it seemed to make sense;

etc.

No tests were deleted or altered in any significant way (except when a test was a clear duplicate of another one).