fix: Raycast from CircleHitbox's center (#1918)
When casting ray from center of a CircleHitbox, the ray didn't detect intersections with the hitbox. This was happening because of ray origin and absolute center of hitbox being coincident. When this happened, the temporary end point for the line segment used for detecting ray intersection turned out to be (0, 0). This basically meant, the ray length was zero. This PR makes sure that when such case occurs, we use the ray direction as the temporary end point.