fix: OpacityEffect rounding error calculation (#1933)
When using OpacityEffect.fadeOut, the value of opacity at the end of the effect was non zero. I suspect that the rounding error calculation is wrong and this PR tried to simplify that logic.

Proof that this PR actually fixes the problem can be seen at the test results of 793415f and 5f3c591. First one just added a test for OpacityEffect.fadeOut and second added the actual fix.

Disclaimer: I don't fully understand this part of the code, so be extra vigilant while reviewing this PR.