feat: Make _ButtonState public for SpriteButtonComponent (#1941)
This PR makes the internal button state of SpriteButtonComponent public. This will make it easier for users to manually change the button state, if needed, by change the value of current. One use case discussed on discord was to make the button appear visually pressed on press on a key.