refactor: move broadphase-related functionality into separate subdir (#1943)
A variant of refactoring mentioned here #1894 (comment)
Sweep broadphase have no any additional specific classes, it's based only on core classes. So it is discussion question, do we need to keep int in separate folder (to avoid mixing with core broadphase.dart file) or just leave it in the root folder 'broadphase'.