feat: Allow flame_forge2d's followBodyComponent to follow centre of mass (#1947)
This PR adds the optional named parameter useCenterOfMass to followBodyComponent.
It defaults to false so no breaking changes.
Setting true allows camera following that does not "wobble" when an otherwise stationary BodyComponent (with an offset centre of mass) rotates.

Also removed reference to component anchor in documentation/comments since BodyComponent does not use an anchor.