test: Windows specific test failures (#1975)
This PR modifies the Vector2.scaleTo test to use expectDouble instead of expect so as to account for the minor deviation in floating point math.

For the second test, elapsed time of sprite animation was reported as 0.0 after pump() in some cases, causing the test to fail. Even on Linux the elapsed value turns out to be quite small but never reaches zero. For some reason the animation gets reset quite a few time for each pump, causing the elapsed time to get reset to 0. This PR changes the test to check if the animation has started instead of checking is elapsed is non-zero.