feat: Lexer for the flame_yarn project (#2081)
This PR adds the new flame_yarn repository (see #2066), and implements the tokenizer (lexer) for the .yarn file format.

By itself, the tokenizer is an internal function. However, it is a well-defined component that can be developed and tested standalone. Therefore, I am creating a separate PR for it -- otherwise tokenizer+lexer would have been unmanageably large.