Removed gets in ShapeComponent constructor.
Removed unnecessary get of paintLayers (and paint) even if null on constructor. This would have caused HasPaint to create a paintLayers instance when no paintLayers passed in contructor.
Added more HasPaint tests to cover cases when both paint and paintLayers passed on constructor and when paintLayers is cleared or set to [] by the user.