fix: Focus handling with a scope on the `GameWidget` (#1725)
This ignores keyboard events if the game widget doesn't have the primary focus. This allows focus nodes on overlay widgets to take precedence on keyboard resolving.

Also, add focus scope in the game widget for when the overlay is removed, the focus goes back to the game.

Also, add tests for GameWidget for all these cases and some more.