feat: Add `maintainState` property to Route (#2161)
Add a boolean property to the Route component named maintainState. When true (by default), the current behavior remains where the state of the child Component is retained. If maintainState is set to false, the child Component of the Route will be rebuilt when the Route is pushed onto the stack again.