docs: BodyComponents should not be nested (#2242)
Forge2D/Box2D doesn't have a concept of nested bodies and all bodies should live on a top level in the physics world. However, it's hard to know this for people who stat with Forge2d right away. It would be nice to have it clearly mentioned in the documentations.

Closes #2240