fix: Only use initialized game for tests and remove setMount from onGameResize (#2246)
Since we used setMounted in onGameResize the mounted feature didn't work properly when invoked from for example onLoad. This PR makes sure that all tests passes without using setMounted in onLoad.

Some GameWidget tests I want a second opinion on, or ideas of how to make those tests better, because when setting mounted explicitly it kind of cheats on part of those tests.