fix: Jenny now always stringifies whole numbers without .0 (#2265)
The YarnSpinner language supports variables of type numeric, but not integers or doubles. The closest concept in Dart is the num type, which is a union of int and double. However, we need to make sure that a number stringifies in a consistent manner regardless of whether its underlying representation is integer or double. Thus, we must make sure that 12 and 12.0 not only compare equal (which they already do), but also have the same representation when converted to string.

This PR ensures that any integer-like double looks like an integer when converted to a string.