fix: Override `remove()` method to fix the functionality issue in the `FlameMultiBlocProvider` (#2280)
We overrode the add() method in the FlameMultiBlocProvider to handle a new functionality (add the new component to the last provider). We need to apply this functionality to the remove() method too. I just overrode the remove() method to apply that functionality.
In this way, both of add() and remove() methods work consistently.