fix: Camera no longer "sticks" to boundary with BoundedPositionBehavior (#2307)
Previously, when using BoundedPositionBehavior, the camera would "stick" to the world boundary as soon as the target moves outside. With this PR, the camera will now try to keep as close to the target as possible, even if it moves outside of the world boundary. Thus, the camera will now behave as if the boundary became "slippery".

Also added method nearestPoint() in class Shape.