docs: Rename caveace asset to cave_ace in our examples (#2304)
I am on a crusade to improve our spellchecking. I have a big PR that I am now going to break down into very small chunks to for review.

Here I am renaming an asset to include a word separator, as the combined word "caveace" doesn't exist. This not only allows us to remove one unclear entry in our dictionary, but make our example code easier to read for new users.

Note: in order to easily update the internal name of the atlas (unrelated to file name), that is stored in the binary, I used this pending feature from Fire Atlas.