docs: Update cspell github action and configuration (#2325)
I am on a crusade to improve our spellchecking. I have a big PR that I am now going to break down into very small chunks to for review.

zwaldowski's cspell-action has been deprecated in favor of streetsidesoftware/cspell-action.

This PR:

updates to use the new recommended cspell action
update to use checkout v3, because why not
fixes a few violations the new action caught
Note: I think we should make a distinction between things that should be in a dictionary (e.g. AABB, a standard industry acronym that we expect people to know what it is) and specific edge-cases/ignore rules for specific parts of specific files (e.g. something like fccc which is meaningless outside of the very specific context in which it is used). For the latter, I believe we should use case-by case exceptions.