fix: Incorrect JoystickComponent position in landscape mode #2387 (#2389)
When in landscape mode, a Component will need to be repositioned based on the new canvas
dimensions. If a Component has not reached the mounted state when
HudMarginComponent.onGameResize is called, then _updateMargins will not be called, and
the position will not be updated.
This PR removes the check that the Component has reached the mounted state in
HudMarginComponent.onGameResize so that _updateMargins will always be called.