fix: TiledComponent now can be safely loaded regardless of the order (#2391)
Basically, the TiledAtlas must not be reused without clone() because it includes batch which is different for each TiledAtlas.
But the following snippets(before this PR) can lead to the reuse of the TiledAtlas if I have loaded the map with the same tilesets(same atlasKey).

   return atlasMap[key] ??= TiledAtlas._(...); 
Also, this PR includes the change of the TestAssetBundle.
I don't know about the exact intention of it.
But it can load only a single map and also it ignored the actual file name to the bundle's first file name.
This makes tests to be hard to debug.

This change is quite far from the original one. So please feel free to tell me to divide it to the separate PR.