docs: Fix broken image link on flame_tiled pub (#2407)
Fix broken image link on flame_tiled pub.
Sadly the repo relative link works on GitHub but not on pub.
This replaces it with an equivalent githubusercontent URL.
This is how it is done by other packages and recommended on Stack Overflow according to my research (but open to other better solutions).

image

I also made the image a bit bigger bc it was small and hard to see (to me).