fix: TextBoxComponent rendering for new line (#2413)
Initially the calculation of lines and updating accordingly ignores the
case for a new line escape characters.
This PR changes how we calculate and update lines when checking for
bounds.

**Before:**
When a text that contains `\n` character is given to the TextBox for
example `Hello \n World`. It doesn't consider the new line and just
appends the `\n` to the same line. This increases the lineHeight for
that particular word (and won't be updated in this case). But, instead
when there is a text `\nHello World`, then the initial lineHeight itself
is twice than what it actually is.

So, the main problem here is that we are not considering (ignoring) the
`\n` characters.

This can be solved by checking for any new lines in the split word (From
`split(' ')`) by using `split('\n')` and adding the lines appropriately
into the `_lines`.

**After**:
Now, whenever a `\n` is encountered, we respect both the new line and
the spacing right after the `\n` character. This eliminates the improper
rendering of the text. Also, `lineHeight` will always be same as
expected and not more.